pyproject.toml
[project]
name = "TinyGen"
version = "0.1.0"
description = "Default template for PDM package"
authors = [{ name = "Ellen Agarwal", email = "ellenagarwal897@gmail.com" }]
dependencies = [
  "fastapi>=0.110.1",
  "uvicorn>=0.29.0",
  "gitpython>=3.1.43",
  "openai>=1.17.0",
  "loguru>=0.7.2",
  "supabase>=2.4.2",
  "python-dotenv>=1.0.1",
]
requires-python = ">=3.11.0"
readme = "README.md"
license = { text = "MIT" }

[tool.pdm.scripts]
dev = "uvicorn tinygen:app --reload"
prod = "uvicorn tinygen:app --host 0.0.0.0"

[tool.pdm.dev-dependencies]
dev = ["pytest>=8.1.1", "httpx>=0.27.0"]
